Rename application.js to appbase.js (#4080)
