+++ src/deprecated/deprecated.js
@@ -95,7 +95,7 @@ import { TouchDevice } from '../input/touch-device.js';
 import { getTouchTargetCoords, Touch, TouchEvent } from '../input/touch-event.js';
 
 import { FILLMODE_FILL_WINDOW, FILLMODE_KEEP_ASPECT, FILLMODE_NONE, RESOLUTION_AUTO, RESOLUTION_FIXED } from '../framework/constants.js';
-import { Application } from '../framework/application.js';
+import { Application } from '../framework/app-base.js';
 import { getApplication } from '../framework/globals.js';
 import { CameraComponent } from '../framework/components/camera/component.js';
 import { Component } from '../framework/components/component.js';
