+++ src/framework/components/joint/system.js
@@ -4,7 +4,7 @@ import { ComponentSystem } from '../system.js';
 import { JointComponent } from './component.js';
 import { JointComponentData } from './data.js';
 
-/** @typedef {import('../../application.js').Application} Application */
+/** @typedef {import('../../app-base.js').Application} Application */
 
 const _schema = ['enabled'];
