+++ src/framework/components/light/system.js
@@ -9,7 +9,7 @@ import { ComponentSystem } from '../system.js';
 import { _lightProps, LightComponent } from './component.js';
 import { LightComponentData } from './data.js';
 
-/** @typedef {import('../../application.js').Application} Application */
+/** @typedef {import('../../app-base.js').Application} Application */
 
 const lightTypes = {
     'directional': LIGHTTYPE_DIRECTIONAL,
