+++ src/framework/components/screen/system.js
@@ -8,7 +8,7 @@ import { ComponentSystem } from '../system.js';
 import { ScreenComponent } from './component.js';
 import { ScreenComponentData } from './data.js';
 
-/** @typedef {import('../../application.js').Application} Application */
+/** @typedef {import('../../app-base.js').Application} Application */
 
 const _schema = ['enabled'];
