+++ src/framework/components/sound/system.js
@@ -9,7 +9,7 @@ import { SoundComponent } from './component.js';
 import { SoundComponentData } from './data.js';
 
 /** @typedef {import('../../../sound/manager.js').SoundManager} SoundManager */
-/** @typedef {import('../../application.js').Application} Application */
+/** @typedef {import('../../app-base.js').Application} Application */
 
 const _schema = ['enabled'];
