+++ src/index.js
@@ -210,7 +210,7 @@ export { getTouchTargetCoords, Touch, TouchEvent } from './input/touch-event.js'
 // FRAMEWORK
 export * from './framework/constants.js';
 export { script } from './framework/script.js';
-export { app, Application } from './framework/application.js';
+export { app, Application } from './framework/app-base.js';
 export { AnimationComponent } from './framework/components/animation/component.js';
 export { AnimationComponentSystem } from './framework/components/animation/system.js';
 export * from './anim/controller/constants.js';
