+++ src/vr/vr-manager.js
@@ -2,7 +2,7 @@ import { EventHandler } from '../core/event-handler.js';
 
 import { VrDisplay } from './vr-display.js';
 
-/** @typedef {import('../framework/application.js').Application} Application */
+/** @typedef {import('../framework/app-base.js').Application} Application */
 
 /**
  * Manage and update {@link VrDisplay}s that are attached to this device.
