+++ test/asset/asset.test.mjs
@@ -1,4 +1,4 @@
-import { Application } from '../../src/framework/application.js';
+import { Application } from '../../src/framework/app-base.js';
 import { Asset } from '../../src/asset/asset.js';
 
 import { HTMLCanvasElement } from '@playcanvas/canvas-mock';
