+++ test/framework/application.test.mjs
@@ -1,5 +1,5 @@
 import { FILLMODE_KEEP_ASPECT, RESOLUTION_FIXED } from '../../src/framework/constants.js';
-import { Application } from '../../src/framework/application.js';
+import { Application } from '../../src/framework/app-base.js';
 import { AssetRegistry } from '../../src/asset/asset-registry.js';
 import { BatchManager } from '../../src/scene/batching/batch-manager.js';
 import { ComponentSystemRegistry } from '../../src/framework/components/registry.js';
